Revert "Merge branch 'main' of https://github.com/AnimusindustriesOSX/Pirate-Platformer"
This reverts commit d25e12b3fb408f6fe3721d25e905ed0afc917a16, reversing
changes made to eb8804c804f74afd3e247825144e6cd99b2ea634.

+++ Assets/Scripts/UserInterfaceController.cs
@@ -8,7 +8,7 @@ public class UserInterfaceController : MonoBehaviour
 {
     // Start is called before the first frame update
     [SerializeField] float health = 100;
-    [SerializeField] float insanity;
+    [SerializeField] int insanity;
 
     [SerializeField] int maxHealth = 100;
     [SerializeField] int maxInsanity = 100;
